+++ client/package.json
@@ -12,10 +12,12 @@
     "cra-build-watch": "^3.2.0",
     "formik": "^2.1.5",
     "framer-motion": "^2.7.5",
+    "history": "^5.0.0",
     "react": "^16.12.0",
     "react-cool-portal": "^1.0.18",
     "react-dom": "^16.12.0",
     "react-responsive": "^8.1.0",
+    "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "^3.4.3",
     "react-use-clipboard": "^1.0.1",
     "reconnecting-websocket": "^4.4.0",
