+++ client/src/Game/index.js
@@ -1,2 +1,1 @@
 export { default as Container } from "./Container";
-export { default as reducer } from "./reducer";
