+++ client/src/index.js
@@ -1,18 +1,18 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { BrowserRouter } from "react-router-dom";
 
 import "./index.css";
 import { SocketProvider } from "./SocketContext";
 import { SoundProvider } from "./Sounds/Context";
-import { StoreProvider } from "./Store";
 import { Container as GameContainer } from "./Game";
 
 ReactDOM.render(
   <SocketProvider>
     <SoundProvider>
-      <StoreProvider>
+      <BrowserRouter>
         <GameContainer />
-      </StoreProvider>
+      </BrowserRouter>
     </SoundProvider>
   </SocketProvider>,
   document.getElementById("root")
